# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para interagir com o Redirectify",
        "title": "API do Redirectify",
        "contact": {
            "name": "Equipe da DevOps (Pablo Eduardo, Guilherme Bernardo e Eduardo Henrique)",
            "email": "comercialdevops@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/redirecionadores/:codigo_hash/link/:id": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Links"
                ],
                "summary": "Remove um link específico de um redirecionador específico",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Código Hash",
                        "name": "codigo_hash",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Links"
                ],
                "summary": "Atualiza um link específico de um redirecionador específico",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Código Hash",
                        "name": "codigo_hash",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Nome",
                        "name": "nome",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Link",
                        "name": "link",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Plataforma",
                        "name": "plataforma",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    }
                }
            }
        },
        "/redirecionadores/:codigo_hash/links": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Links"
                ],
                "summary": "Retorna os links do redirecionador com o código hash fornecido",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Código Hash",
                        "name": "codigo_hash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Link"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Links"
                ],
                "summary": "Cria links no redirecionador com o código hash fornecido",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Código Hash",
                        "name": "codigo_hash",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Links",
                        "name": "links",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.LinkToBatchInsert"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    }
                }
            }
        },
        "/redirecionadores/:codigo_hash/links/:id": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Links"
                ],
                "summary": "Retorna o link com o id fornecido do redirecionador com o código hash fornecido",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Código Hash",
                        "name": "codigo_hash",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Link"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Erro"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Erro": {
            "type": "object",
            "properties": {
                "message": {}
            }
        },
...
        "Link": {
            "type": "object",
            "properties": {
                "atualizado_em": {
                    "type": "string"
                },
                "criado_em": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "link": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "plataforma": {
                    "type": "string"
                },
                "redirecionador": {
                    "type": "string"
                },
                "removido_em": {
                    "type": "string"
                }
            }
        },
...
        "models.LinkToBatchInsert": {
            "type": "object",
            "properties": {
                "link": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "plataforma": {
                    "type": "string"
                }
            }
        },
...
    }
}